Using TSID to avoid round-trip to generate id

+++ src/main/java/com/victor/labs/tms_mass_updates/dto/ReservaRequestDTO.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class ReservaRequestDTO {
     private FiltroDTO filtro;
-    private Integer planejamentoId;
+    private Long planejamentoId;
 }
