+++ src/main/java/com/victor/labs/tms_mass_updates/repository/DocumentoCargaRepository.java
@@ -3,5 +3,5 @@ package com.victor.labs.tms_mass_updates.repository;
 import com.victor.labs.tms_mass_updates.domain.DocumentoCarga;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DocumentoCargaRepository extends JpaRepository<DocumentoCarga, Integer> {
+public interface DocumentoCargaRepository extends JpaRepository<DocumentoCarga, Long> {
 }
