+++ src/main/java/com/victor/labs/tms_mass_updates/repository/PlanejamentoRepository.java
@@ -3,5 +3,5 @@ package com.victor.labs.tms_mass_updates.repository;
 import com.victor.labs.tms_mass_updates.domain.Planejamento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlanejamentoRepository extends JpaRepository<Planejamento, Integer> {
+public interface PlanejamentoRepository extends JpaRepository<Planejamento, Long> {
 }
